attribute restructure and diagonal bullet movement

+++ src/main/java/com/bham/bc/components/BackendServices.java
@@ -47,9 +47,9 @@ public interface BackendServices {
      */
     void removeObstacle(GenericObstacle obstacle);
 
-    int getPlayerX();
+    double getPlayerX();
 
-    int getPlayerY();
+    double getPlayerY();
 
     Shape getHomeHitBox();
