+++ src/main/java/com/bham/bc/entity/physics/BombTank.java
@@ -27,7 +27,7 @@ public class BombTank extends BaseGameEntity {
 	 * @param x coordinate of x axis
 	 * @param y coordinate of y axis
 	 */
-	public BombTank(int x, int y) {
+	public BombTank(double x, double y) {
 		super(GetNextValidID(), x, y);
 		initImages();
 	}
